+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Next
 
+- fix: do not clear events when reset is called ([#170](https://github.com/PostHog/posthog-android/pull/170))
+- fix: reload feature flags as anon user after reset is called ([#170](https://github.com/PostHog/posthog-android/pull/170))
+
 ## 3.6.0 - 2024-08-29
 
 - recording: expose session id ([#166](https://github.com/PostHog/posthog-android/pull/166))